modification et ajout fonctionnalité

+++ src/pages/adminInclude/user/listeUser.php
@@ -2,16 +2,25 @@
 
     if (isset($_SESSION['user']['role']) && $_SESSION['user']['role'] == 'admin') {
         if (isset($_GET['deleteUser']) && isset($_GET['deleteUser']) == true) {
-            $delGameId = htmlspecialchars($_GET['value']);
+            $delUserId = htmlspecialchars($_GET['value']);
 
-            if (isset($_GET['avatar'])) {
+            if (isset($_GET['avatar']) && $_GET['avatar'] != '../../src/img/site/defaut_avatar.png') {
                 unlink($_GET['avatar']);
             }
 
-            deleteUser($delGameId);
+            deleteUser($delUserId);
             header('location: ../../src/pages/admin.php?choix=listeUser');
             exit();
         }
+
+        if (isset($_GET['banUser']) && isset($_GET['banUser']) == true) {
+            $banUserId = htmlspecialchars($_GET['value']);
+
+            banUser($banUserId);
+            header('location: ../../src/pages/admin.php?choix=listeUser');
+            exit();
+        }
+
     }
     $listeUser = getUser();
 
@@ -32,6 +41,7 @@
                 <td>Email</td>
                 <td>Role</td>
                 <td>Supprimer</td>
+                <td>Bannir</td>
             </tr>
             <?php 
 
@@ -45,6 +55,18 @@
                     <td><?=$value['email']?></td>
                     <td><?=$value['nomRole']?></td>
                     <td class="ta-c tc-r"><a href="../../src/pages/admin.php?choix=listeUser&deleteUser=true&value=<?=$value['userId']?>&avatar=<?=$value['avatar']?>"><i class="far fa-plus-square"></i></a></td>
+                    <?php 
+                        if ($value['ban'] == null) {
+                        ?>
+                            <td class="ta-c tc-r"><abbr title="L'utilisateur est déjà banni" style="color:red;"><i class="fas fa-ban"></i></abbr></td>
+                        <?php
+                        } else {
+                        ?>
+                            <td class="ta-c tc-r"><a href="../../src/pages/admin.php?choix=listeUser&banUser=true&value=<?=$value['userId']?>"><i class="fas fa-user-slash"></i></a></td>
+                        <?php
+                        }
+                    ?>
+                    
                 </tr>
             <?php
                 }
